test(davinci): satisfy impeto quality gates
