+++ crates/vize_impeto/src/op/id.rs
@@ -1,0 +1,76 @@
+use core::fmt;
+
+/// One operation id in a [`super::Program`].
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
+pub struct OpId(u32);
+
+impl OpId {
+    /// Create an operation id from its numeric index.
+    #[must_use]
+    pub const fn new(index: u32) -> Self {
+        Self(index)
+    }
+
+    /// Numeric index.
+    #[must_use]
+    pub const fn index(self) -> u32 {
+        self.0
+    }
+}
+
+impl fmt::Display for OpId {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "op#{}", self.0)
+    }
+}
+
+/// One region id. Region `0` is the required root region.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
+pub struct RegionId(u32);
+
+impl RegionId {
+    /// The root region every program must define.
+    pub const ROOT: Self = Self(0);
+
+    /// Create a region id from its numeric index.
+    #[must_use]
+    pub const fn new(index: u32) -> Self {
+        Self(index)
+    }
+
+    /// Numeric index.
+    #[must_use]
+    pub const fn index(self) -> u32 {
+        self.0
+    }
+}
+
+impl fmt::Display for RegionId {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "r#{}", self.0)
+    }
+}
+
+/// One effect-scope id.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
+pub struct EffectId(u32);
+
+impl EffectId {
+    /// Create an effect-scope id from its numeric index.
+    #[must_use]
+    pub const fn new(index: u32) -> Self {
+        Self(index)
+    }
+
+    /// Numeric index.
+    #[must_use]
+    pub const fn index(self) -> u32 {
+        self.0
+    }
+}
+
+impl fmt::Display for EffectId {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "fx#{}", self.0)
+    }
+}
