+++ crates/vize_impeto/tests/op_family.rs
@@ -3,27 +3,27 @@ use vize_impeto::op::{EdgeKind, OpKind, Phase};
 #[test]
 fn op_kind_mnemonics_are_the_vapor_generalization_set() {
     let kinds = [
-        OpKind::SetProp,
-        OpKind::SetDynamicProps,
-        OpKind::SetText,
-        OpKind::SetEvent,
-        OpKind::SetHtml,
-        OpKind::SetTemplateRef,
-        OpKind::InsertNode,
-        OpKind::PrependNode,
-        OpKind::Directive,
-        OpKind::If,
-        OpKind::For,
-        OpKind::CreateComponent,
-        OpKind::SlotOutlet,
-        OpKind::GetTextChild,
-        OpKind::ChildRef,
-        OpKind::NextRef,
+        (OpKind::SetProp, "impeto.set-prop"),
+        (OpKind::SetDynamicProps, "impeto.set-dynamic-props"),
+        (OpKind::SetText, "impeto.set-text"),
+        (OpKind::SetEvent, "impeto.set-event"),
+        (OpKind::SetHtml, "impeto.set-html"),
+        (OpKind::SetTemplateRef, "impeto.set-template-ref"),
+        (OpKind::InsertNode, "impeto.insert-node"),
+        (OpKind::PrependNode, "impeto.prepend-node"),
+        (OpKind::Directive, "impeto.directive"),
+        (OpKind::If, "impeto.if"),
+        (OpKind::For, "impeto.for"),
+        (OpKind::CreateComponent, "impeto.create-component"),
+        (OpKind::SlotOutlet, "impeto.slot-outlet"),
+        (OpKind::GetTextChild, "impeto.get-text-child"),
+        (OpKind::ChildRef, "impeto.child-ref"),
+        (OpKind::NextRef, "impeto.next-ref"),
     ];
     assert_eq!(kinds.len(), 16);
-    for kind in kinds {
-        assert_eq!(OpKind::from_mnemonic(kind.mnemonic()), Some(kind));
-        assert!(kind.mnemonic().starts_with("impeto."));
+    for (kind, mnemonic) in kinds {
+        assert_eq!(kind.mnemonic(), mnemonic);
+        assert_eq!(OpKind::from_mnemonic(mnemonic), Some(kind));
     }
 }
