+++ davinci-road/plan/folio-format.md
@@ -337,38 +337,6 @@ exact oracles in `tests/verifier_observer.rs`.
 
 ## Impeto page (P3-1)
 
-The S3 stage dump is `vize_impeto::folio::S3Folio`, a derived Folio page named
-`[s3-folio]`. It is flat rather than tree-shaped: regions, ops, state edges,
-and effect scopes are separate order-bearing sections whose id references are
-validated by the TS-27 phase validator.
-
-Header:
-
-| field   | values                                 | notes                    |
-| ------- | -------------------------------------- | ------------------------ |
-| `phase` | `built`, `partitioned`, or `scheduled` | printed from `op::Phase` |
-
-Sections, in struct declaration order:
-
-| section              | entry grammar                                                                   |
-| -------------------- | ------------------------------------------------------------------------------- |
-| `[s3-folio.regions]` | `id=<n> parent=<n\|-> owner=<n\|-> span=<start>:<end>`                          |
-| `[s3-folio.ops]`     | `id=<n> kind=<mnemonic> region=<n> effect=<n\|-> span=<start>:<end>`            |
-| `[s3-folio.edges]`   | `from=<n> to=<n> kind=<dom-order\|effect-order\|data-dependency> effect=<n\|->` |
-| `[s3-folio.effects]` | `id=<n> owner=<n> region=<n> span=<start>:<end>`                                |
-
-The optional id value is spelled `-` when absent. The op mnemonic set is the
-P3-1 generalization of Vapor's OperationNode family:
-`impeto.set-prop`, `impeto.set-dynamic-props`, `impeto.set-text`,
-`impeto.set-event`, `impeto.set-html`, `impeto.set-template-ref`,
-`impeto.insert-node`, `impeto.prepend-node`, `impeto.directive`, `impeto.if`,
-`impeto.for`, `impeto.create-component`, `impeto.slot-outlet`,
-`impeto.get-text-child`, `impeto.child-ref`, and `impeto.next-ref`.
-
-`Display` is identical to `Full` because this page is generated by
-`#[derive(Folio)]`; semantic elision remains a hand-written-page choice. The
-tests in `crates/vize_impeto/tests/folio_laws.rs` pin
-`print(parse(t)) == t` on canonical text and structural `parse(print(v)) == v`.
-TS-27 (`crates/vize_impeto/tests/phase_validator.rs`) separately pins exact
-rejections for unresolved state edges, malformed region nesting, effect-scope
-leaks, and scheduled-phase back edges.
+The S3 stage dump is documented in
+[`folio-format-impeto.md`](./folio-format-impeto.md). That page owns the flat
+Impeto grammar and the TS-27 validator contract.
