+++ tests/tooling/davinci-stage-dependencies.test.ts
@@ -210,7 +210,7 @@ test("Davinci DOM production imports lowering through the physical S1-to-S2 pack
   }
 });
 
-test("Davinci atelier core S2 witnesses import lowering through the physical S1-to-S2 package", () => {
+test("Atelier core S2 witnesses import lowering through the physical S1-to-S2 package", () => {
   const lowering = dependency(metadata, "vize_atelier_core", "vize_s1_to_s2", "dev");
   assert.equal(lowering.rename, null);
   const dependencies = workspacePackage(metadata, "vize_atelier_core").dependencies;
@@ -234,53 +234,27 @@ test("Davinci atelier core S2 witnesses import lowering through the physical S1-
   }
 });
 
-test("Vize CLI package imports S0 storage through the stage alias", () => {
-  assertS0AliasConsumer({
-    packageName: "vize",
-    label: "vize package",
-    directory: path.join(repoRoot, "crates", "vize"),
-  });
-});
+const s0AliasConsumers = [
+  ["vize", "vize package", ["crates", "vize"]],
+  ["vize_test_runner", "Test runner", ["tests", "vize_test_runner"]],
+  ["vize_armature", "Armature parser", ["crates", "vize_armature"]],
+  ["vize_patina", "Patina linter", ["crates", "vize_patina"]],
+  ["vize_musea", "Musea component gallery", ["crates", "vize_musea"]],
+  ["vize_fresco", "Fresco TUI", ["crates", "vize_fresco"]],
+  ["vize_curator", "Curator reporting utilities", ["crates", "vize_curator"]],
+  ["vize_vitrine", "Vitrine bindings", ["crates", "vize_vitrine"]],
+  ["vize_maestro", "Maestro LSP", ["crates", "vize_maestro"]],
+  ["vize_atelier_ssr", "Atelier SSR compiler", ["crates", "vize_atelier_ssr"]],
+  ["vize_atelier_dom", "Atelier DOM compiler", ["crates", "vize_atelier_dom"]],
+  ["vize_atelier_jsx", "Atelier JSX compiler", ["crates", "vize_atelier_jsx"]],
+  ["vize_relief", "Relief AST", ["crates", "vize_relief"]],
+] as const;
 
-test("Test runner imports S0 storage through the stage alias", () => {
-  assertS0AliasConsumer({
-    packageName: "vize_test_runner",
-    label: "Test runner",
-    directory: path.join(repoRoot, "tests", "vize_test_runner"),
+for (const [packageName, label, parts] of s0AliasConsumers) {
+  test(`${label} imports S0 storage through the stage alias`, () => {
+    assertS0AliasConsumer({ packageName, label, directory: path.join(repoRoot, ...parts) });
   });
-});
-
-test("Armature parser imports S0 storage through the stage alias", () => {
-  assertS0AliasConsumer({
-    packageName: "vize_armature",
-    label: "Armature parser",
-    directory: path.join(repoRoot, "crates", "vize_armature"),
-  });
-});
-
-test("Patina linter imports S0 storage through the stage alias", () => {
-  assertS0AliasConsumer({
-    packageName: "vize_patina",
-    label: "Patina linter",
-    directory: path.join(repoRoot, "crates", "vize_patina"),
-  });
-});
-
-test("Musea component gallery imports S0 storage through the stage alias", () => {
-  assertS0AliasConsumer({
-    packageName: "vize_musea",
-    label: "Musea component gallery",
-    directory: path.join(repoRoot, "crates", "vize_musea"),
-  });
-});
-
-test("Fresco TUI imports S0 storage through the stage alias", () => {
-  assertS0AliasConsumer({
-    packageName: "vize_fresco",
-    label: "Fresco TUI",
-    directory: path.join(repoRoot, "crates", "vize_fresco"),
-  });
-});
+}
 
 test("Canon content-mapper imports S0 storage through the stage alias", () => {
   const manifest = readRepoFile("crates", "vize_canon", "Cargo.toml");
@@ -291,54 +265,6 @@ test("Canon content-mapper imports S0 storage through the stage alias", () => {
     label: "Canon content-mapper",
     directory: path.join(repoRoot, "crates", "vize_canon", "src", "batch", "virtual_project"),
     filter: (fullPath) => path.basename(fullPath).startsWith("content_mapper"),
-  });
-});
-
-test("Curator reporting utilities import S0 storage through the stage alias", () => {
-  assertS0AliasConsumer({
-    packageName: "vize_curator",
-    label: "Curator reporting utilities",
-    directory: path.join(repoRoot, "crates", "vize_curator"),
-  });
-});
-
-test("Vitrine bindings import S0 storage through the stage alias", () => {
-  assertS0AliasConsumer({
-    packageName: "vize_vitrine",
-    label: "Vitrine bindings",
-    directory: path.join(repoRoot, "crates", "vize_vitrine"),
-  });
-});
-
-test("Maestro LSP imports S0 storage through the stage alias", () => {
-  assertS0AliasConsumer({
-    packageName: "vize_maestro",
-    label: "Maestro LSP",
-    directory: path.join(repoRoot, "crates", "vize_maestro"),
-  });
-});
-
-test("Atelier SSR compiler imports S0 storage through the stage alias", () => {
-  assertS0AliasConsumer({
-    packageName: "vize_atelier_ssr",
-    label: "Atelier SSR compiler",
-    directory: path.join(repoRoot, "crates", "vize_atelier_ssr"),
-  });
-});
-
-test("Atelier DOM compiler imports S0 storage through the stage alias", () => {
-  assertS0AliasConsumer({
-    packageName: "vize_atelier_dom",
-    label: "Atelier DOM compiler",
-    directory: path.join(repoRoot, "crates", "vize_atelier_dom"),
-  });
-});
-
-test("Atelier JSX compiler imports S0 storage through the stage alias", () => {
-  assertS0AliasConsumer({
-    packageName: "vize_atelier_jsx",
-    label: "Atelier JSX compiler",
-    directory: path.join(repoRoot, "crates", "vize_atelier_jsx"),
   });
 });
 
@@ -356,12 +282,4 @@ test("Atelier core compiler macros import S0 storage through the stage alias", (
     const source = readRepoFile("crates", "vize_atelier_core", "src", relative);
     assert.doesNotMatch(source, /\bvize_carton\b/u, `${relative} must use vize_s0 or $crate`);
   }
-});
-
-test("Relief AST imports S0 storage through the stage alias", () => {
-  assertS0AliasConsumer({
-    packageName: "vize_relief",
-    label: "Relief AST",
-    directory: path.join(repoRoot, "crates", "vize_relief"),
-  });
 });
